Repairs to the BigNum divide logic.
Various parts of this were assuming behavior in vu_sub that was
contrary to the actual interface contract. This change makes the
computations slightly more expensive, but fixes those errors. Also a
minor bug fix in asString().

I am not completely convinced that this was the best fix -- it might
have been better to change the vu_sub contract.