Rotate [io]BinaryStream to use conventional streams.
This is an incompatible change. I revised the [io]BinaryStream classes
to be built on top of conventional streams, and dropped the
[io]fBinaryStream classes. The main benefit of this change is that we
can now use gzstream classes for mkimage and the like.

An unfortunate consequence of this change is that the library license
had to change to BSD/LGPL according to what parts are used by the user.