Update to crt1.S to correct a bad JBSR instruction.
This is why SpaceBank was faulting -- there was an extra indirection
in this instruction. With this change made, I get past the fault and
subsequently hit a problem in memcpy related to coldfire instruction
selection. The latter appears to be a newlib build issue, and I'm
working on that now..