LexLoc now generalized from path to arbitrary named origin.
This is a backwards compatibel change. The old form still works, but
the preferred use is now to provide an origin string rather than a
Path pointer.

I need to build and push new RPMS, and I can't do that until morning,
but I can at least get the commit pushed.