Hack the boost_base script to follow the convention implied by $libdir.
Problem: On 64-bit systems, the old script would fail because it
checked /usr/lib rather than /usr/lib64. Worse, on systems where
dual-mode libraries were installed it would mysteriously succeed.

This patch changes the behavior when the generic --with-libdir is
provided to configure. If the basename of libdir is 'mumble', then the
probe strategy is altered to use 'mumble' in place of 'lib'. This at
least resolves the problem for 64-bit RPM-based systems. I don't know
any general solution.